splitting excel output into file destination and in-memory stream
